Fixed index bug, added server icons
Added react-advanced-cropper (very early stage but looks really promising).
Previously setCurrentChannel on createChannel was called at the end and it didn't work due to 1. not being based on the array length (it sets to the latest created server) and 2. since it had to be based on array length, it would be better used inside the map.
Fixed key issues when iterating through chat (it was based on the name only)