Added authentication (middleware too)
Merged http server with sockets to handle authentication with as little as socket power as possible (will use it to also get the servers a user is already in).
Need to restructure User object, remove password and keep user (and maybe email) for local storage purposes.
Need to check for authentication on '/chat' navegation to load all servers and such on connect.