Fixed authentication
Now returns servers that the server is in.
Big bug/unintended behavior, the credentials/user that is logged in isn't changed per client connected, so the client that logs in after another gets all the servers. Creating a server after transmits it to the previous client.
Chat messages now correctly links links.